Documents/ClaudeCodeTesting/requirements.txt
watchdog
psutil
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pystray
Pillow
